requirements.txt
pyyaml
click
pycryptodome
